add pop-up cat add correct work residual

+++ carts/models.py
@@ -30,7 +30,7 @@ class Cart(models.Model):
     select_buy = models.BooleanField(default=True, verbose_name='VX')
     user = models.ForeignKey(to=User, on_delete=models.CASCADE, blank=True, null=True, verbose_name='Пользователь')
     product = models.ForeignKey(to=Products, on_delete=models.CASCADE, verbose_name='Товар')
-    quantity = models.PositiveSmallIntegerField(default=0, verbose_name='Количество')
+    quantity = models.DecimalField(default=0.00, max_digits=7, decimal_places=2, verbose_name='Количество')
     session_key = models.CharField(max_length=32, blank=True, null=True)
     created_timestamp = models.DateTimeField(auto_now_add=True, verbose_name='Дата добавления')
